+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added support for custom types derived from primitive types in `database/sql` driver
 * Bumped `golang.org/x/net` from 0.35.0 to 0.38.0
 
 ## v3.125.0